+++ src/Engine/SpellFxRenderer.cpp
@@ -456,18 +456,13 @@ void SpellFxRenderer::_4A7948_mind_blast_after_effect(SpriteObject* a1)
 }
 
 //----- (004A7A27) --------------------------------------------------------
-bool SpellFxRenderer::AddMobileLight(SpriteObject* a1, unsigned int uDiffuse,
-	int uRadius)
+bool SpellFxRenderer::AddMobileLight(SpriteObject* a1, unsigned int uDiffuse, int uRadius)
 {
-	return pMobileLightsStack->AddLight(a1->vPosition.toFloat(), a1->uSectorID,
-		uRadius, uDiffuse & 0x000000FF, (uDiffuse & 0x0000FF00) >> 8,
-		(uDiffuse & 0x00FF0000) >> 16, _4E94D3_light_type);
+	return pMobileLightsStack->AddLight(a1->vPosition, a1->uSectorID, uRadius, uDiffuse & 0x000000FF, (uDiffuse & 0x0000FF00) >> 8, (uDiffuse & 0x00FF0000) >> 16, _4E94D3_light_type);
 }
 
 //----- (004A7A66) --------------------------------------------------------
-void SpellFxRenderer::
-_4A7A66_miltiple_spell_collision_partifles___like_after_sparks_or_lightning(
-	SpriteObject* a1, unsigned int uDiffuse, Texture* texture, float a4)
+void SpellFxRenderer::_4A7A66_miltiple_spell_collision_partifles___like_after_sparks_or_lightning(SpriteObject* a1, unsigned int uDiffuse, Texture* texture, float a4)
 {
 	int v5;               // eax@1
 	double v7;            // st6@1
